requirements.txt
Django==3.2.14
djangorestframework==3.13.1
psycopg2==2.9.3
PyJWT==2.3.0